Connections: move the CanTmcl derived classes in a sub-directory
