+++ pytrinamic/connections/__init__.py
@@ -1,10 +1,10 @@
 from .dummy_tmcl_interface import DummyTmclInterface
-from .pcan_tmcl_interface import PcanTmclInterface
-from .socketcan_tmcl_interface import SocketcanTmclInterface
-from .kvaser_tmcl_interface import KvaserTmclInterface
+from .can_tmcl.pcan_tmcl_interface import PcanTmclInterface
+from .can_tmcl.socketcan_tmcl_interface import SocketcanTmclInterface
+from .can_tmcl.kvaser_tmcl_interface import KvaserTmclInterface
 from .serial_tmcl_interface import SerialTmclInterface
 from .uart_ic_interface import UartIcInterface
 from .usb_tmcl_interface import UsbTmclInterface
-from .slcan_tmcl_interface import SlcanTmclInterface
-from .ixxat_tmcl_interface import IxxatTmclInterface
+from .can_tmcl.slcan_tmcl_interface import SlcanTmclInterface
+from .can_tmcl.ixxat_tmcl_interface import IxxatTmclInterface
 from .connection_manager import ConnectionManager
